requirements.txt
APScheduler==3.7.0
Pillow
aiohttp
asyncio
bs4
dnspython
emoji
gitpython
google-search==1.1.1
google==3.0.0
googlesearch-python==1.0.1
gpytranslate
gtts
heroku3
humanize
langs
lottie
motor
natsort
numpy
prettytable
psutil
psycopg2-binary
ptable
py-Ayiin==0.4.6
pybase64
pymediainfo
pymongo[srv]
python-dotenv
pyyaml
rarfile
redis
requests
speedtest-cli==2.1.3
sqlalchemy==1.3.23
telegraph
textblob
tgcrypto
tracemoepy
opencv-python-headless
urllib3
uvloop
wget==3.2
yins-music==0.0.7.dev3
youtube-search-python
yt-dlp
zipfile38
